+++ packages/pie-toolbox/src/code/math-rendering-accessible/mathjax-script.js
@@ -1,4 +1,42 @@
-export const initializeMathJax = () => {
+import { getGlobal } from './render-math';
+
+export const initializeMathJax = (renderOpts) => {
+  if (renderOpts?.useSingleDollar) {
+    // eslint-disable-next-line
+    console.warn('[math-rendering] using $ is not advisable, please use $$..$$ or \\(...\\)');
+  }
+
+  const texConfig = renderOpts?.useSingleDollar
+    ? {
+        macros: {
+          parallelogram: '\\lower.2em{\\Huge\\unicode{x25B1}}',
+          overarc: '\\overparen',
+          napprox: '\\not\\approx',
+          longdiv: '\\enclose{longdiv}',
+        },
+        inlineMath: [
+          ['$', '$'],
+          ['\\(', '\\)'],
+        ],
+        processEscapes: true,
+        displayMath: [
+          ['$$', '$$'],
+          ['\\[', '\\]'],
+        ],
+      }
+    : {
+        macros: {
+          parallelogram: '\\lower.2em{\\Huge\\unicode{x25B1}}',
+          overarc: '\\overparen',
+          napprox: '\\not\\approx',
+          longdiv: '\\enclose{longdiv}',
+        },
+        displayMath: [
+          ['$$', '$$'],
+          ['\\[', '\\]'],
+        ],
+      };
+
   // Create a new promise that resolves when MathJax is ready
   window.mathjaxLoadedP = new Promise((resolve) => {
     // Set up the MathJax configuration
@@ -14,29 +52,16 @@ export const initializeMathJax = () => {
             mathjax.handleRetriesFor(() => rerender.call(this, start));
           };
           MathJax.startup.defaultReady();
+          // Set the MathJax instance in the global object
+          const globalObj = getGlobal();
+          globalObj.instance = MathJax;
           resolve();
         },
       },
       loader: {
         load: ['input/mml'],
       },
-      tex: {
-        inlineMath: [
-          ['$', '$'],
-          ['\\(', '\\)'],
-        ],
-        macros: {
-          parallelogram: '\\lower.2em{\\Huge\\unicode{x25B1}}',
-          overarc: '\\overparen',
-          napprox: '\\not\\approx',
-          longdiv: '\\enclose{longdiv}',
-        },
-        displayMath: [
-          ['$$', '$$'],
-          ['\\[', '\\]'],
-        ],
-        processEscapes: true,
-      },
+      tex: texConfig,
       chtml: {
         fontURL: 'https://unpkg.com/mathjax-full@3.2.2/ts/output/chtml/fonts/tex-woff-v2',
         displayAlign: 'center',
